Debounce sliders

+++ tool/obsm_calculator.py
@@ -141,7 +141,7 @@ class Spell:
     def _calc_derived_fields(self):
         self.dominant_school = self._determine_school()
         self.total_cost = self._calc_cost()
-        self.gold_cost = self.total_cost * 3
+        self.gold_cost = int(self.total_cost * 3)
         self.skill_required = self._determine_skill_req()
 
     def _determine_school(self):
